+++ MiamIOSFramework/MiamIOSFramework/ui/Catalog/CatalogView.swift
@@ -27,21 +27,23 @@ public struct CatalogView: View {
 
     @SwiftUI.State private var headerHeight = 50.0
     
+    let closeCatalogAction: (() -> Void)?
     private var usesPreferences = false
     
-    public init(usesPreferences: Bool = false) {
+    public init(usesPreferences: Bool = false , closeCatalogAction: (() -> Void)? = nil) {
         self.catalog = CatalogVM()
         self.usesPreferences = usesPreferences
+        self.closeCatalogAction = closeCatalogAction
     }
     
-    public init(categoryId: String, title: String, usesPreferences: Bool = false) {
+    public init(categoryId: String, title: String, usesPreferences: Bool = false, closeCatalogAction: (() -> Void)? = nil) {
         self.catalog = CatalogVM(categoryID: categoryId, title: title)
-        self.usesPreferences = usesPreferences
+        self.closeCatalogAction = closeCatalogAction
     }
 
     public var body: some View {
         VStack(alignment: .center, spacing: 0.0) {
-            CatalogViewHeader()
+            CatalogViewHeader(closeCatalogAction: closeCatalogAction)
                 .frame(height: catalog.content == .categories ? 60.0 : 0.0)
 
             CatalogToolbarView(showBackButton: (catalog.content != .categories),
@@ -200,9 +202,10 @@ internal struct CatalogPackageRow: View {
 
 @available(iOS 14, *)
 internal struct CatalogViewHeader: View {
+    let closeCatalogAction: (() -> Void)?
     var body: some View {
         if (Template.sharedInstance.catalogViewHeaderTemplate != nil) {
-            Template.sharedInstance.catalogViewHeaderTemplate!
+            Template.sharedInstance.catalogViewHeaderTemplate!(closeCatalogAction)
         } else {
             HStack {
                 Image.miamImage(icon: .ideeRepas)